Fixed bug 10. Domain now created with bounce by default and now you can modify the catchall again.
It uses an older version of the file, so it won't work with the new Kloxo without merging the two, but this version will work with old versions of Kloxo.

The reason is that the new version of Kloxo has a more secure method of unzipping the file that unzips it under the actual client's user account, rather than the root user. The old Kloxo does not do that.

I recommend all of you download this to your servers, and Danny, if you would like, you could push this out in an update, as it is old-kloxo friendly, new-kloxo unfriendly.